+++ teflo/constants.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright (C) 2021 Red Hat, Inc.
+# Copyright (C) 2022 Red Hat, Inc.
 #
 #    This program is free software: you can redistribute it and/or modify
 #    it under the terms of the GNU General Public License as published by
@@ -20,7 +20,7 @@
 
     A module containing all constants used throughout the teflo code base.
 
-    :copyright: (c) 2021 Red Hat, Inc.
+    :copyright: (c) 2022 Red Hat, Inc.
     :license: GPLv3, see LICENSE for more details.
 
 """
@@ -180,3 +180,7 @@ LOGGING_CONFIG = {
                        'propagate': False},
     }
 }
+
+# Default retries for installing ansible dependencies
+ANSIBLE_GALAXY_INSTALL_DELAY = 30
+ANSIBLE_GALAXY_INSTALL_ATTEMPTS = 2