+++ teflo/teflo.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright (C) 2021 Red Hat, Inc.
+# Copyright (C) 2022 Red Hat, Inc.
 #
 #    This program is free software: you can redistribute it and/or modify
 #    it under the terms of the GNU General Public License as published by
@@ -20,7 +20,7 @@
 
     Teflo a framework to test product interoperability.
 
-    :copyright: (c) 2021 Red Hat, Inc.
+    :copyright: (c) 2022 Red Hat, Inc.
     :license: GPLv3, see LICENSE for more details.
 """
 from copy import deepcopy
@@ -215,6 +215,8 @@ class Teflo(LoggerMixin, TimeMixin):
                            'TEFLO_RESULTS_FOLDER': os.path.abspath(self.config.get('RESULTS_FOLDER')),
                            'TEFLO_WORKSPACE': os.path.abspath(self.config.get('WORKSPACE'))
                            })
+
+        self.current_task = None
 
     @property
     def name(self):
@@ -513,11 +515,12 @@ class Teflo(LoggerMixin, TimeMixin):
             self.logger.info("." * 50)
         # initialize overall status
         status = 0
-        if not self.teflo_options.get('no_notify', False):
-            self.notify('on_start', status, passed_tasks, failed_tasks, scenario=sc)
-
         try:
             for task in sort_tasklist(tasklist):
+                self.current_task = task
+                if not self.teflo_options.get('no_notify', False):
+                    self.notify('on_start', status, passed_tasks, failed_tasks, scenario=sc)
+
                 self.logger.info(' * Task    : %s' % task)
 
                 # initially update list of passed tasks
@@ -688,7 +691,7 @@ class Teflo(LoggerMixin, TimeMixin):
         data = {}
 
         # create a pipeline builder object
-        pipe_builder = PipelineFactory.get_pipeline(task)
+        pipe_builder = PipelineFactory.get_pipeline(task, self.current_task)
 
         # check if teflo supports the task
         if not pipe_builder.is_task_valid():