+++ teflo/utils/pipeline.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright (C) 2021 Red Hat, Inc.
+# Copyright (C) 2022 Red Hat, Inc.
 #
 #    This program is free software: you can redistribute it and/or modify
 #    it under the terms of the GNU General Public License as published by
@@ -22,7 +22,7 @@
     Module containing classes and functions for building pipelines of resources
     and tasks for teflo to run.
 
-    :copyright: (c) 2021 Red Hat, Inc.
+    :copyright: (c) 2022 Red Hat, Inc.
     :license: GPLv3, see LICENSE for more details.
 """
 
@@ -41,11 +41,21 @@ from ..tasks import CleanupTask
 class PipelineFactory(object):
 
     @staticmethod
-    def get_pipeline(task):
+    def get_pipeline(task, current_task=None):
+        """This method used by regular pipeline and also the notification pipeline.
+        It checks the task name, to which of the objects it belongs to
+        and return the object accordingly.
+
+        :param task: task name
+        :type task: str
+        :param current_task: The current run task from the scenario tasks. Used only for notification pipeline.
+        :type current_task: str
+        :return: pipeline object
+        """
         if task in TASKLIST:
             return PipelineBuilder(task)
         else:
-            return NotificationPipelineBuilder(task)
+            return NotificationPipelineBuilder(task, current_task)
 
 
 class PipelineBuilder(object):
@@ -233,10 +243,11 @@ class NotificationPipelineBuilder(PipelineBuilder):
     The primary class for building teflos notification pipelines to execute.
     """
 
-    def __init__(self, trigger):
+    def __init__(self, trigger, current_task):
 
         super(NotificationPipelineBuilder, self).__init__('notify')
         self.trigger = trigger
+        self.current_task = current_task
 
     def is_task_valid(self):
         """Check if the pipeline task name is valid for teflo.
@@ -295,7 +306,8 @@ class NotificationPipelineBuilder(PipelineBuilder):
             scenario_notifications = [item for item in
                                         filter_notifications_on_trigger(self.trigger, scenario_notifications,
                                                                         getattr(scenario, 'passed_tasks'),
-                                                                        getattr(scenario, 'failed_tasks'))
+                                                                        getattr(scenario, 'failed_tasks'),
+                                                                        self.current_task)
                                       ]
 
             # notification resource