+++ teflo/utils/resource_checker.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright (C) 2021 Red Hat, Inc.
+# Copyright (C) 2022 Red Hat, Inc.
 #
 #    This program is free software: you can redistribute it and/or modify
 #    it under the terms of the GNU General Public License as published by
@@ -22,7 +22,7 @@
     Module containing classes and functions for  validating services and other resources before running the teflo
     scenario
 
-    :copyright: (c) 2021 Red Hat, Inc.
+    :copyright: (c) 2022 Red Hat, Inc.
     :license: GPLv3, see LICENSE for more details.
 """
 
@@ -82,14 +82,7 @@ class ResourceChecker(object):
                 self.ans_service.galaxy_options = item.get('ansible_galaxy_options', None)
 
                 # downloading ansible_roles
-                try:
-                    self.ans_service.download_roles()
-                except (TefloError, AnsibleServiceError):
-                    if 'retry' in self.ans_service.galaxy_options and self.ans_service.galaxy_options['retry']:
-                        LOG.info("Download failed.  Sleeping 5 seconds and \
-                                          trying again")
-                        time.sleep(5)
-                        self.ans_service.download_roles()
+                self.ans_service.download_roles()
 
                 error_msg = ''
                 try: